refactor(assess): decompose test_pressure into a package and sharpen two heuristics
Split the 869-LOC, ccn-216 test_pressure.py into a focused package behind the
existing scan_test_pressure() facade, with no change to the public API or the
test_pressure output schema:

- common.py     - shared file walking and test-file detection
- mutation.py   - mutation-config detection, bounded opt-in runs, aggregation
- heuristics.py - the three cheap detectors
- aggregate.py  - TestPressureResult schema + scan_test_pressure orchestration
- __init__.py   - stable facade re-exporting every former name

Sharpen two cheap heuristics that produced false positives on this repo:

- assertion_on_internal: a private attribute in call position (mod._helper(...))
  is the subject under test, not an internal-state read. Excluding it removes 8
  false positives (direct unit tests of private helpers) while keeping the
  planted hollow-fixture true positive (a private-field read).
- duplicate_truth: restrict to fields - instance/class attributes and
  module/class-level names - excluding function-local bare-name aliasing
  (raw = result.stdout). Removes all 14 false positives.

Closes #78
Closes #81
Closes #82
